refactor DtoMetadata.kt into restapi/global/
