+++ client/android/app/src/main/java/com/sju18001/petmanagement/restapi/dto/AccountDto.kt
@@ -1,6 +1,6 @@
 package com.sju18001.petmanagement.restapi.dto
 
-import com.sju18001.petmanagement.restapi.dao.DtoMetadata
+import com.sju18001.petmanagement.restapi.global.DtoMetadata
 
 data class LoginReqDto (
     val username: String,
