--- client/android/app/src/main/java/com/sju18001/petmanagement/restapi/dto/DtoMetadata.kt
@@ -1,7 +1,0 @@
-package com.sju18001.petmanagement.restapi.dto
-
-data class DtoMetadata (
-    val status: Boolean,
-    val message: String,
-    val exception: String
-)
